Implemented DelegatedDispatchVerifier for CennznetDispatchVerifier
